requirements.txt
gymnasium<0.29.0
transforms3d>=0.4.1
numpy==1.23.5
dm-control<1.1.0
glfw==2.5.6
gym==0.23.1
imageio==2.25.1
imageio-ffmpeg==0.4.8
tensorboard==2.11.2
tensorflow==2.11.0
tensorflow-probability==0.19.0
jax==0.3.25
jaxlib==0.3.25 -f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
flax==0.5.3
natsort==8.2.0
opencv-python<5.0.0
pandas==2.0.0
pygame==2.3.0
tqdm==4.64.1
wandb<0.15.0
shimmy>=0.2.1
scipy < 2.0.0
chex==0.1.6
mujoco==2.3.3
mujoco_utils==0.0.5
pyserde[msgpack]==0.10.7
psutil==5.9.5
scikit-image
filterpy==1.4.5
numba
ml_collections>=0.1.0
optax==0.1.4
absl-py==1.4.0
moviepy
dmcgym